# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Сервис анализа патентной активности компаний.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/filters": {
      "post": {
        "tags": [
          "Filters"
        ],
        "summary": "Create Filter",
        "description": "Создает новый фильтр на основе загруженного Excel-файла.\n\nArgs:\n    name (str): Имя нового фильтра.\n    file (UploadFile): Загружаемый Excel-файл с данными для фильтра.\n    session (AsyncSession): Сессия для взаимодействия с базой данных.\n\nRaises:\n    HTTPException(400): Если формат загруженного файла не поддерживается.\n    HTTPException(500): Если произошла ошибка при создании фильтра.\n\nReturns:\n    FilterDB: Созданный фильтр.",
        "operationId": "create_filter_filters_post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_filter_filters_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Filters"
        ],
        "summary": "Read Filters",
        "description": "Получает список всех фильтров.\n\nArgs:\n    session (AsyncSession): Сессия для взаимодействия с базой данных.\n\nReturns:\n    List[FilterDB]: Список всех фильтров.",
        "operationId": "read_filters_filters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FilterDB"
                  },
                  "title": "Response Read Filters Filters Get"
                }
              }
            }
          }
        }
      }
    },
    "/filters/{filter_id}": {
      "get": {
        "tags": [
          "Filters"
        ],
        "summary": "Read Filter",
        "description": "Получает фильтр по его идентификатору.\n\nArgs:\n    filter_id (int): Идентификатор фильтра.\n    session (AsyncSession): Сессия для взаимодействия с базой данных.\n\nRaises:\n    HTTPException(404): Если фильтр с указанным идентификатором не найден.\n\nReturns:\n    FilterDB: Фильтр с указанным идентификатором.",
        "operationId": "read_filter_filters__filter_id__get",
        "parameters": [
          {
            "name": "filter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Filter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Filters"
        ],
        "summary": "Update Filter",
        "description": "Обновляет имя существующего фильтра.\n\nArgs:\n    filter_id (int): Идентификатор фильтра.\n    name (str): Новое имя фильтра.\n    session (AsyncSession): Сессия для взаимодействия с базой данных.\n\nReturns:\n    FilterDB: Обновленный фильтр.",
        "operationId": "update_filter_filters__filter_id__put",
        "parameters": [
          {
            "name": "filter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Filter Id"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Filters"
        ],
        "summary": "Delete Filter",
        "description": "Удаляет существующий фильтр.\n\nArgs:\n    filter_id (int): Идентификатор фильтра.\n    session (AsyncSession): Сессия для взаимодействия с базой данных.\n\nReturns:\n    FilterDB: Удаленный фильтр.",
        "operationId": "delete_filter_filters__filter_id__delete",
        "parameters": [
          {
            "name": "filter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Filter Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_filter_filters_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_filter_filters_post"
      },
...
      "FilterDB": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created"
          },
          "tax_numbers_count": {
            "type": "integer",
            "title": "Tax Numbers Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "filename",
          "id",
          "created",
          "tax_numbers_count"
        ],
        "title": "FilterDB"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}